Changed default route to login
Added index as key to each comment.

This is not recommended and each comment should have a unique iq. But
since our comments are not going to be deleted/edited its okay to use
for now.

Added author name to avatar to comment list

Vertical nav bar and newsfeed appearance changes

aesthetic changes to navbar color